refactor: move message sync logic to application layer
- Extract HandleRealtimeMessage from SDK to app-level Syncer
- SDK now only handles transport, emits raw messages
- Add local persistence store for messages
- Add debug log panel with sliding window buffer
- Add conversationID format support (g: for groups)
- Unify PullMessages parameter types (int64)

+++ sdks/go/social/service.go
@@ -12,7 +12,7 @@ type RosterClient interface {
 	ListConversations(ctx context.Context, accessToken string) ([]types.Conversation, error)
 	GetConversationMaxSeqs(ctx context.Context, accessToken string) ([]types.ReadSnapshot, error)
 	GetConversationReadSnapshots(ctx context.Context, accessToken string) ([]types.ReadSnapshot, error)
-	PullMessages(ctx context.Context, accessToken string, ranges []types.SeqRange, limitPerConversation int) ([]types.PulledConversationMessages, error)
+	PullMessages(ctx context.Context, accessToken string, ranges []types.SeqRange, limitPerConversation int64) ([]types.PulledConversationMessages, error)
 	AckReadSeq(ctx context.Context, accessToken, conversationID string, readSeq int64) error
 }
 
@@ -64,7 +64,7 @@ func (s *RosterService) LoadInitialData(ctx context.Context, accessToken string)
 	}, nil
 }
 
-func (s *RosterService) PullMessages(ctx context.Context, accessToken string, ranges []types.SeqRange, limit int) ([]types.PulledConversationMessages, error) {
+func (s *RosterService) PullMessages(ctx context.Context, accessToken string, ranges []types.SeqRange, limit int64) ([]types.PulledConversationMessages, error) {
 	return s.client.PullMessages(ctx, accessToken, ranges, limit)
 }
